fix get_args

+++ get_args.py
@@ -12,7 +12,7 @@ def get_args():
     parser.add_argument('--traj_length', type=int, default=10, help='trajectory length')
     parser.add_argument('--dim', type=int, default=1, help='dimension of hidden state')
     parser.add_argument('--loss_type', type=str, default='forward_kl', help='type of KL regularization')
-    parser.add_argument('--ent_coef', type=int, default=0.1, help='coefficient on KL regularization term (loss_type)')
+    parser.add_argument('--ent_coef', type=float, default=0.1, help='coefficient on KL regularization term (loss_type)')
     parser.add_argument('--subroutine', type=str, default='train_agent', help='what subroutine to run (train_agent, evaluate_agent, ess_dim, ess_traj)')
 
     # parse
